Silence unreachable pattern warnings
When a feature is complete, the "not implemented" branches are never
reached. But for DX, being able to just implement the parser, compiler,
or bytecode instruction on its own without having to implement
everythiing before compiling for the first time is very helpful. A
readable error is shown at runtime in either case.